♻️ Refactor code & add comments
