+++ src/components/3D/Ground.js
@@ -1,8 +1,5 @@
 import { usePlane } from '@react-three/cannon';
-import { useLayoutEffect, useRef } from 'react';
-import * as THREE from 'three';
-import Line from './Line';
-import colors from './assets/styles/scss/abstracts/variables.scss';
+import colors from '../../assets/styles/scss/abstracts/variables.scss';
 
 export const Ground = ({ stepSize }) => {
   const [ref] = usePlane(() => ({ rotation: [-Math.PI / 2, 0, 0] }));
