+++ src/utils/imageFromPainting.js
@@ -1,0 +1,8 @@
+/**
+ *
+ * @param {*} painting Object
+ * @returns ImageObject
+ */
+const imageFromPainting = (painting) => painting.images.overall.images[0];
+
+export default imageFromPainting;
